fix: move review api functions
